+++ pineboolib/main.py
@@ -26,6 +26,7 @@ import pineboolib
 from pineboolib import qt3ui
 from pineboolib.dbschema.schemaupdater import parseTable
 from pineboolib.qsaglobals import aqtt
+from pineboolib.flcontrols import FLForm
 import pineboolib.emptyscript
 
 from pineboolib import qsatype, qsaglobals, DlgConnect
@@ -476,45 +477,6 @@ class XMLAction(XMLStruct):
     def unknowSlot(self):
         print("Executing unknow script for Action", self.name)
         #Aquí debería arramcar el script
-
-class FLForm(QtGui.QWidget):
-    known_instances = {}
-    def __init__(self, parent, action, load=False):
-        try:
-            assert (self.__class__,action) not in self.known_instances
-        except AssertionError:
-            print("WARN: Clase %r ya estaba instanciada, reescribiendo!. " % ((self.__class__,action),)
-                + "Puede que se estén perdiendo datos!" )
-        self.known_instances[(self.__class__,action)] = self
-        QtGui.QWidget.__init__(self, parent)
-        self.action = action
-        self.prj = action.prj
-        self.mod = action.mod
-        self.layout = QtGui.QVBoxLayout()
-        self.layout.setMargin(2)
-        self.layout.setSpacing(2)
-        self.setLayout(self.layout)
-        # self.widget = QtGui.QWidget()
-        # self.layout.addWidget(self.widget)
-        self.bottomToolbar = QtGui.QFrame()
-        self.bottomToolbar.setMaximumHeight(64)
-        self.bottomToolbar.setMinimumHeight(16)
-        self.bottomToolbar.layout = QtGui.QHBoxLayout()
-        self.bottomToolbar.setLayout(self.bottomToolbar.layout)
-        self.bottomToolbar.layout.setMargin(0)
-        self.bottomToolbar.layout.setSpacing(0)
-        self.bottomToolbar.layout.addStretch()
-        self.toolButtonClose = QtGui.QToolButton()
-        self.toolButtonClose.setIcon(QtGui.QIcon(filedir("icons","gtk-cancel.png")))
-        self.toolButtonClose.clicked.connect(self.close)
-        self.bottomToolbar.layout.addWidget(self.toolButtonClose)
-        self.layout.addWidget(self.bottomToolbar)
-        self.setWindowTitle(action.alias)
-        self.loaded = False
-        if load: self.load()
-
-    def load(self):
-        if self.loaded: return
 
 class FLMainForm(FLForm):
     """ Controlador dedicado a las ventanas maestras de búsqueda (en pestaña) """
